Turn Suspension into an interface
This is equivalent to the interface change in https://github.com/reactphp/async/pull/15 to allow different implementations.

This also allows decorating `Suspension` to implement listeners like proposed in https://github.com/revoltphp/event-loop/pull/2.
